requirements.txt
future
emoji==1.7.0
lxml
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==13.14
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp
Pillow
CurrencyConverter
#googletrans
google_trans_new
jikanpy
speedtest-cli
coffeehouse
regex
bleach
markdown2
wikipedia
telethon
spamwatch
alphabet_detector
pyrate-limiter
covid
datetime
pytz
cachetools
ujson
bs4
fontTools
hachoir
pybase64
pySmartDL
validators
gtts
pretty_errors
tswift
PyYAML
humanize
pytube
